backend/requirements.txt
google-cloud-aiplatform
fastapi[standard]
python-dotenv
firebase-admin